LPD-59867 baseline
# breaking

## What /modules/apps/object/object-admin-rest-impl/rest-openapi.yaml

A new query parameter to postObjectDefinition and putObjectDefinitionByExternalReferenceCode.

## Why

To introduce a switch in validation behavior by using a parameter on these endpoints.

----

# breaking

## What /modules/apps/object/object-service/src/main/java/com/liferay/object/service/impl/ObjectDefinitionLocalServiceImpl.java

Added accumulateError and objectFields parameters.
Added a new thread local to manage the accumulation and throwing of validation exceptions. Also, to delegate the new thread local for accumulating or throwing validation exceptions.

## Why

To make possible to implement the validation accumulation strategy

----